+++ modules/canvaskit/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - Skottie accessors for dynamic text properties (text string, font size).
+
 ## [0.25.0] - 2021-03-02
 
 ### Added